+++ src/SubMapper/Implementation/SubMapper.php
@@ -16,6 +16,7 @@ namespace Rekalogika\Mapper\SubMapper\Implementation;
 use Rekalogika\Mapper\Context\Context;
 use Rekalogika\Mapper\Exception\UnexpectedValueException;
 use Rekalogika\Mapper\ObjectCache\ObjectCache;
+use Rekalogika\Mapper\SubMapper\Exception\CacheNotSupportedException;
 use Rekalogika\Mapper\SubMapper\SubMapperInterface;
 use Rekalogika\Mapper\Transformer\Contracts\MainTransformerAwareInterface;
 use Rekalogika\Mapper\Transformer\Contracts\MainTransformerAwareTrait;
@@ -36,7 +37,7 @@ class SubMapper implements SubMapperInterface, MainTransformerAwareInterface
         private PropertyTypeExtractorInterface $propertyTypeExtractor,
         private PropertyAccessorInterface $propertyAccessor,
         private mixed $source,
-        private Type $targetType,
+        private ?Type $targetType,
         private Context $context,
     ) {
     }
@@ -122,6 +123,10 @@ class SubMapper implements SubMapperInterface, MainTransformerAwareInterface
 
     public function cache(mixed $target): void
     {
+        if ($this->targetType === null) {
+            throw new CacheNotSupportedException($this->context);
+        }
+
         ($this->context)(ObjectCache::class)->saveTarget(
             source: $this->source,
             targetType: $this->targetType,