+++ src/SubMapper/SubMapperFactoryInterface.php
@@ -25,7 +25,7 @@ interface SubMapperFactoryInterface
     public function createSubMapper(
         MainTransformerInterface $mainTransformer,
         mixed $source,
-        Type $targetType,
+        ?Type $targetType,
         Context $context
     ): SubMapperInterface;
 }